script： 修复 本地部署测试文件，未检测目录导致copy异常

+++ scripts/localTest.js
@@ -1,7 +1,7 @@
 /*
  * @Author: 2460392754@qq.com
  * @Date: 2022-03-02 00:43:54
- * @LastEditTime: 2022-03-02 11:41:28
+ * @LastEditTime: 2022-03-03 15:58:57
  * @Description: 本地部署测试文件，替换 `examples` 里各项目的 `node_modules/**\/dist` 文件内容
  * @FilePath: /uniapp-router-view-loader/scripts/localTest.js
  */
@@ -21,8 +21,10 @@ for (const childPath of dirChildrenPathAll) {
     const toDir = './' + path.join(dirPath, childPath, 'node_modules', name, 'dist')
 
     try {
-        fse.copySync('./dist', toDir)
-        console.log(chalk.yellow(`Tip: 测试项目本地部署成功【${toDir}】`));
+        if (fs.existsSync(toDir)) {
+            fse.copySync('./dist', toDir)
+            console.log(chalk.yellow(`Tip: 测试项目本地部署成功【${toDir}】`));
+        }
     } catch (err) {
         console.error(err)
     }
